Replace FieldValue with ProtoFieldValue
